Dodanie testów jednostkowych do repository

+++ quiz_mode/src/main/java/com/rafalskrzypczyk/quiz_mode/presentation/question_details/ui_models/SimpleCategoryUIModel.kt
@@ -10,6 +10,6 @@ data class SimpleCategoryUIModel(
 fun Category.toPresentation() : SimpleCategoryUIModel {
     return SimpleCategoryUIModel(
         name = title,
-        color = color
+        color = color.toLong()
     )
 }
